Remove fetching the repository
Doing the fetching ourselves means we need to either forward a lot of
options to the `actions/checkout` action or make this action very
inflexible.

The main point of this action is to provide some common configuration
that all repositories will use by default.

Also update the token input name to make it more clear this is for
GitHub and to match the protolint action name.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>